Move to Global_Fields namespace, structure like blocks
